fix: type int to double

+++ src/FontesAppCalculadora/Calculadora.java
@@ -4,31 +4,31 @@ package FontesAppCalculadora;
 public class Calculadora{
 
       // atributo
-      private int resultado = 0;
+      private double resultado = 0;
 
       // método somar
-      public int somar( int n1, int n2 ){
+      public double somar( double n1, double n2 ){
 
           resultado = n1 + n2;
           return resultado;
       }
 
       // método subtrair
-      public int subtrair(int n1, int n2 ){
+      public double subtrair(double n1, double n2 ){
 
           resultado = n1 - n2;
           return resultado;
       }        
 
       // método multiplicar
-      public int multiplicar( int n1, int n2 ){
+      public double multiplicar( double n1, double n2 ){
 
           resultado = n1 * n2;
           return resultado;
       }
 
       // método dividir
-      public int dividir( int n1, int n2 ){
+      public double dividir( double n1, double n2 ){
 
           resultado = n1 / n2;
           return resultado;
